requirements.txt
chardet==3.0.4
coloredlogs==15.0.1
cryptography>=41.0.6
edgegrid-python==1.3.1
emojis==0.7.0
prettytable==0.7.2
pyyaml==6.0.1
requests==2.31.0
rich_argparse==1.4.0
tabulate==0.9.0
urllib3>=1.26.18
xlsxwriter==1.2.8
